modify #21 Validationの実行を外部クラスに移動
